Add extra helper function to import export
